Switch from certificate to public key
Previously the client had a trusted certificate by the server, which it used to  verify the identity of the peer. However in the handshake only the public key of the certificate was ever used. From now on instead of a certificate, only the public key is saved by the client.

The baseserver now generates the public key from it's private key at runtime, meaning the public key doesn't need to be passed into the baseserver  class