Switch from own encryption to TLSv1.3
Switched to TLSv1.3, which is infinitely safer than any own encryption algorithm.

Added some useful decorators

Moved ping from a built in base server method to it's own function in network.py. This is because it writes a command to a client, meaning the client has to be configured to receive this specific command, which is probably why it would not be in the base server class