Improved accepting multiple clients
The server used to only accept clients using one thread, meaning at worst, a single client could hang the accept thread for 30ish seconds on the default 10sec timeout. Now we use a thread pool to accept multiple clients as they come in.

Additionally clients used to get accepted using a while loop and a timeout, now a select is used to wait for a client to arrive and then sending the connection to the thread pool